refactor(core): извлечь ядро insightCards (CRUD/reorder/decide/vault/коллекции)
